Remove pkg-config dependency.
Used to use this in extconf.rb, but there's no way to tell it to pass
the --static flag for pkg-config, so it's not really worthwhile.

This still depends on pkg-config existing on Windows, so who knows how
well that'll pan out. Granted, this isn't tested on Windows or Linux
because I have neither.