Implement glfwGetGammaRamp/glfwSetGammaRamp bindings
Get returns a hash of red, green, and blue arrays. The keys are, of
course, :red, :green, and :blue. This seemed like the most reasonable
approach to their implementation, so I went ahead with that.

Set takes the same input - you provide a hash of the same format with
arrays of values, it will pass those onto GLFW via the gamma ramp
struct. Fairly simple. A good way to test this is to do something like

  ramp = (0 ... 32768).step(128).to_a
  monitor.gamma_ramp = { red: ramp, green: ramp, blue: ramp }

which should give the monitor a half-bright linear gamma ramp. It'll
probably be more interesting to do something useful with it, but it's an
easy way to see if it works.

It's recommended you do stick to using 256-length gamma ramps.